Refactored SendMediaComponent to only handle the file input and start to use the SendMediaService
